Add files via upload
Added new feature to import hour_readings ( (1) If all hour values exist in the DB (from the reading), the new hour_reading will be imported as a new row in the DB but with different column differences to indicate that this hour_reading is different and there is a change based on the old one, (2) If there isn't selected or matched an existing row in the DB the new hour_reading will be imported as a new row in the DB, (3) If there is a match of an old hour_reading in the DB but only some of the hour values exist (e. g. 1:00 exists but 0:00 does not) the hour reading values will be set to the new ones where they are not set).