Add files via upload
Fixed routing bugs - there was a conflict because there were 2 files with the same name but in different folders and the express router middleware was not working properly. Added new pages - Graph STP Clients datatable page and an internal page - Graph STP client for each client that has a profile_id assigned. Inside the page there is a fullcalendar that calculates the load(amount) for each hour in the prediction of the Graph STP.