Add files via upload
Fixed visualization bug with the sidebar menu. Added a new functionality - when the identification_code is clicked in the hour-readings table (menu readings) and in the graph-predictions table (menu graphs) it redirects to clients info page and it opens the calendar for the given type of reading and given day automatically. Refactored code in some files.