Add files via upload
Added new functionality : To calculate imbalances for clients with many different criterias (client name, date from, date to, clientID, ERP type, metering type). It calculates the imbalances and shows a calendar / chart with the result. Added a switch to change between calendar and chart and another switch to change between bar and line type chart. Added css for that page.