Add files via upload
Added preview (handsontable - unique, excel-like datatable) in inquiry menu (readings, graphs and imbalances) of datatable of how information will look like when downloaded. Added a lot of animations. Fixed a bug in the inquiry - graph page (was not getting data correctly). Made changes in the invoicing page - added checkboxes for rows (readings) and col (headings) to issue invoices dynamically. Information is saved in the localstorage and later will be rendered from the preview invoicing page.