Add files via upload
ChartJS now shows data for more than one day correctly. Added realtime onChange filters in the clients menu to search clients by name and ID. Added info labels in some of the listings to show the end user for what period of time is the reading and what is the client ID. Fixed bootstrap switch issue where the CSS selector was not specific enough and did not trigger JS code always.